+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Bug Fixes
 * Fix parents of animated transforms getting deleted. [(more)](https://github.com/d4rkc0d3r/d4rkAvatarOptimizer/issues/14)
+* Always enable SkinnedMeshRenderer component that gets other meshes merged into it. [(more)](https://github.com/d4rkc0d3r/d4rkAvatarOptimizer/issues/15)
 
 ## v1.7
 ### Bug Fixes